refactor: CT+E2Eマージ廃止しE2E単独でカバレッジ閾値 80% を達成
- CT+E2Eのカバレッジマージ方式を廃止しE2E(Chromium)のみで計測する方式に変更
- playwright.config.ts の monocart-reporter に onEnd コールバックで
  閾値チェックを追加（statements/branches/functions/lines 各 80%）
- playwright-ct.config.ts からカバレッジ設定を削除
  （@vis.gl/react-google-maps の Vite エイリアスは維持）
- scripts/merge-coverage.mjs を削除
- .github/workflows/ci.yml から merge ジョブを削除
- package.json から coverage:merge スクリプトを削除
- tests/__mocks__/react-google-maps.tsx の AdvancedMarker が onClick を
  破棄していたのを div にフォワードするよう修正
