+++ .github/workflows/ci.yml
@@ -34,12 +34,6 @@ jobs:
           name: playwright-report-components
           path: playwright-report/
           retention-days: 30
-      - uses: actions/upload-artifact@v4
-        if: ${{ !cancelled() }}
-        with:
-          name: coverage-ct
-          path: coverage-reports/ct/
-          retention-days: 1
 
   build:
     timeout-minutes: 30
@@ -119,7 +113,7 @@ jobs:
           retention-days: 1
 
   coverage:
-    needs: [componentstest, e2etest]
+    needs: [e2etest]
     timeout-minutes: 15
     runs-on: ubuntu-latest
     steps:
@@ -127,27 +121,18 @@ jobs:
       - uses: actions/setup-node@v4
         with:
           node-version: lts/*
-      - name: Install dependencies
-        run: npm install -g pnpm@10 && pnpm install
-      - name: Download CT coverage
-        uses: actions/download-artifact@v4
-        with:
-          name: coverage-ct
-          path: coverage-reports/ct
       - name: Download E2E Chromium coverage
         uses: actions/download-artifact@v4
         with:
           name: coverage-e2e-chromium
           path: coverage-reports/e2e
-      - name: Merge coverage and check thresholds
-        run: pnpm run coverage:merge
       # 注意: このレポートにはソースコードが埋め込まれる。リポジトリを public 化する場合は
       # artifact の公開範囲についても見直すこと
       - uses: actions/upload-artifact@v4
         if: ${{ !cancelled() }}
         with:
-          name: coverage-merged
-          path: coverage-reports/merged/
+          name: coverage-report
+          path: coverage-reports/e2e/
           retention-days: 30
 
   lint:
