+++ package.json
@@ -9,8 +9,7 @@
     "lint": "next lint",
     "test:e2e": "APP_ENV=test playwright test",
     "test:components": "playwright test -c playwright-ct.config.ts",
-    "show-report": "playwright show-report",
-    "coverage:merge": "node scripts/merge-coverage.mjs"
+    "show-report": "playwright show-report"
   },
   "dependencies": {
     "@emotion/cache": "^11.14.0",
